feat: app uses new router
